task 0: update

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -7,7 +7,7 @@
 
 hash_table_t *hash_table_create(unsigned long int size)
 {
-	unsigned int index = 0;
+	unsigned long int index = 0;
 	hash_table_t *table = malloc(sizeof(hash_table_t));
 
 	if (table == NULL)
